This update includes html files. There are several issues with this version that need improving:
(1) Java script at the head of html files could be grouped into a folder for java script only,
(2) html files could be grouped into an independent folder,
(3) the images associated with html files do not automatically compile into the" _images"  folder.

Otherwise, this is a working update. If the radio button html files are not working, then check to see if the images are all in the image folder.