Update magnetics_susceptibility_duplicate.rst
I've made some changes but am a bit lost as to the state of changes.
I've replaced some of the material from the gpg that was omitted. I think the interactive graphic display is effective and we should keep it. The susceptibility graph that you made is informative but so too is the plot that was omitted (adapted from Clark and Emerson). Although there is overlap of certain items, there is some additional information. Let's put both in for now.  There is an inconsistency about hematite on the two plots. 
For magnetic properties of soils some of the statements in the old gpg are relevant and maybe should be included some where. Also, the magnetic susceptibility of rocks often depends upon the % of magnetite. More information about this might be included.