Update induced_polarization_physical_properties_duplicate.rst
I've made some changes but there is major restructuring required to integrate this with other material provided for the Induced Polarization survey. The specification of what IP responses are measured in the lab, and their different units, and if they are time or frequency domain, logically rests here. Then any tables or numbers have some context. 
I'm committing this now but the final restructuring would benefit from a discussion.