Split CI in two Actions: build.yml and test.yml (#234)
Create two different Actions: `test.yml` and `build.yml`. The first one
will run the tests under the `tests` directory that will ensure that the
docs are building properly, that image files have proper extensions and
that links are not broken. The other one will build the docs and deploy
if the event was triggered by a push to `main`. Use more up-to-date
GitHub Actions throughout the workflows and simplify the build and
deployment process. Add a new `requirements-dev.txt` file where we list
dependencies that are needed for running tests, but not to build the
website (like `pytest`). Remove the old GitHub Action workflow
`python-app.yml`.